fix warnings for GCC

+++ semester 6/information-theory/error-correction-codes/serial-emulator/serial_encode.c
@@ -10,7 +10,7 @@ void serial_encode(bits_t *dst, bits_t *src, serial_cfg_t *cfg) {
 	bit b;
 	while (!bitarray_empty(src)) {
 		bitarray_push_back(dst, START_BIT);
-		for (int i = 0; i < cfg->data_bits; ++i) {
+		for (int i = 0; i < (int) cfg->data_bits; ++i) {
 			b = next_bit(src);
 			bitarray_push_back(dst, b);
 			sum += b;
@@ -19,7 +19,7 @@ void serial_encode(bits_t *dst, bits_t *src, serial_cfg_t *cfg) {
 			bitarray_push_back(dst, parity(sum, cfg->parity_bit));
 		}
 		sum = 0;
-		for (int i = 0; i < cfg->stop_bits; ++i) {
+		for (int i = 0; i < (int) cfg->stop_bits; ++i) {
 			bitarray_push_back(dst, STOP_BIT);
 		}
 	}
